moved functionality from processing to business logic
