+++ webviz_subsurface/plugins/_rft_plotter/_callbacks/callbacks.py
@@ -5,8 +5,9 @@ from dash import Dash, Input, Output, State
 from dash.exceptions import PreventUpdate
 
 from .._business_logic import RftPlotterDataModel
+from .._business_logic import filter_frame
 from .._layout import LayoutElements
-from .._processing import filter_frame
+
 from .._figures._crossplot_figure import update_crossplot
 from .._figures._errorplot_figure import update_errorplot
 from .._figures._formation_figure import FormationFigure
